scripts/insert-firestore-doc/requirements.txt
requests
google-auth
google-auth-oauthlib
google-cloud-secret-manager
